Remove scheduling_common_flax and some renames
